+++ src/models/StorageFile.php
@@ -6,6 +6,10 @@ use Yii;
 use luya\web\Application;
 use yii\db\ActiveRecord;
 use luya\helpers\FileHelper;
+use luya\admin\behaviors\LogBehavior;
+use luya\admin\filters\TinyCrop;
+use luya\admin\filters\MediumThumbnail;
+use luya\admin\traits\TaggableTrait;
 
 /**
  * This is the model class for table "admin_storage_file".
@@ -35,6 +39,8 @@ use luya\helpers\FileHelper;
  */
 final class StorageFile extends ActiveRecord
 {
+    use TaggableTrait;
+    
     /**
      * @inheritdoc
      */
@@ -57,6 +63,16 @@ final class StorageFile extends ActiveRecord
     /**
      * @inheritdoc
      */
+    public function behaviors()
+    {
+        return [
+            LogBehavior::class,
+        ];
+    }
+
+    /**
+     * @inheritdoc
+     */
     public static function tableName()
     {
         return 'admin_storage_file';
@@ -67,7 +83,7 @@ final class StorageFile extends ActiveRecord
      */
     public static function find()
     {
-        return parent::find()->orderBy(['name_original' => 'ASC']);
+        return parent::find();
     }
 
     /**
@@ -95,15 +111,17 @@ final class StorageFile extends ActiveRecord
      */
     public function delete()
     {
-        $file = Yii::$app->storage->getFile($this->id);
+        if ($this->beforeDelete()) {
         
-        if ($file && !Yii::$app->storage->fileSystemDeleteFile($file->systemFileName)) {
-            Logger::error("Unable to remove file from filesystem: " . $file->systemFileName);
+            if (!Yii::$app->storage->fileSystemDeleteFile($this->name_new_compound)) {
+                Logger::error("Unable to remove file from filesystem: " . $this->name_new_compound);
+            }
+            
+            $this->updateAttributes(['is_deleted' => true]);
+            
+            $this->afterDelete();
+            return true;
         }
-        
-        $this->updateAttributes(['is_deleted' => true]);
-        
-        return true;
     }
     
     /**
@@ -138,10 +156,129 @@ final class StorageFile extends ActiveRecord
     }
     
     /**
+     * Returns the current file source path for the current filter image.
+     * @return string
+     */
+    public function getSource()
+    {
+        return Yii::$app->storage->fileAbsoluteHttpPath($this->name_new_compound);
+    }
+
+    /**
+     * Get the path to the source files internal, on the servers path.
+     *
+     * This is used when you want to to grab the file on server side for example to read the file
+     * with `file_get_contents` and is the absolut path on the file system on the server.
+     *
+     * @return string The path to the file on the filesystem of the server.
+     * @since 1.2.2.1
+     */
+    public function getServerSource()
+    {
+        return Yii::$app->storage->fileServerPath($this->name_new_compound);
+    }
+
+    /**
+     * Get the size of a file in human readable size.
+     *
+     * For example sizes are partial splitet in readable forms:
+     *
+     * + 100B
+     * + 100KB
+     * + 10MB
+     * + 1GB
+     *
+     * @return string The humand readable size.
+     * @since 1.2.2.1
+     */
+    public function getSizeReadable()
+    {
+        return FileHelper::humanReadableFilesize($this->file_size);
+    }
+
+    /**
+     * Whether the file is of type image or not.
+     *
+     * If the mime type of the files is equals to:
+     *
+     * + `image/gif`
+     * + `image/jpeg`
+     * + `image/jpg`
+     * + `image/png`
+     * + `image/bmp`
+     * + `image/tiff`
+     *
+     * The file indicates to be an image and return value is true.
+     *
+     * @return boolean Whether the file is of type image or not.
+     * @since 1.2.2.1
+     */
+    public function getIsImage()
+    {
+        return in_array($this->mime_type, Yii::$app->storage->imageMimeTypes);
+    }
+
+    /**
+     * Create the thumbnail for this given file if its an image.
+     * 
+     * > This method is used internal when uploading a file which is an image, the file manager preview images are created here.
+     *
+     * @return array Returns an array with the key source which contains the source to the thumbnail.
+     * @since 1.2.2.1
+     */
+    public function getCreateThumbnail()
+    {
+        if (!$this->isImage) {
+            return false;
+        }
+
+        $tinyCrop = Yii::$app->storage->getFilterId(TinyCrop::identifier());
+        foreach ($this->images as $image) {
+            if ($image->filter_id == $tinyCrop) {
+                return ['source' => $image->source];
+            }
+        }
+
+        // create the thumbnail on the fly if not existing
+        $image = Yii::$app->storage->createImage($this->id, $tinyCrop);
+        if ($image) {
+            return ['source' => $image->source];
+        }
+    }
+
+    /**
+     * Create the thumbnail medium for this given file if its an image.
+     * 
+     * > This method is used internal when uploading a file which is an image, the file manager preview images are created here.
+     *
+     * @return array Returns an array with the key source which contains the source to the thumbnail medium.
+     * @since 1.2.2.1
+     */
+    public function getCreateThumbnailMedium()
+    {
+        if (!$this->isImage) {
+            return false;
+        }
+        $mediumThumbnail = Yii::$app->storage->getFilterId(MediumThumbnail::identifier());
+
+        foreach ($this->images as $image) {
+            if ($image->filter_id == $mediumThumbnail) {
+                return ['source' => $image->source];
+            }
+        }
+
+        // create the thumbnail on the fly if not existing
+        $image = Yii::$app->storage->createImage($this->id, $mediumThumbnail);
+        if ($image) {
+            return ['source' => $image->source];
+        }
+    }
+
+    /**
      * @inheritdoc
      */
     public function extraFields()
     {
-        return ['user', 'file', 'images'];
+        return ['user', 'file', 'images', 'source', 'createThumbnail', 'createThumbnailMedium', 'isImage', 'sizeReadable'];
     }
 }