+++ src/storage/BaseFileSystemStorage.php
@@ -224,6 +224,14 @@ abstract class BaseFileSystemStorage extends Component
     ];
 
     /**
+     * @var array a list of mime types which are indicating images
+     * @since 1.2.2.1
+     */
+    public $imageMimeTypes = [
+        'image/gif', 'image/jpeg', 'image/png', 'image/jpg',
+    ];
+
+    /**
      * @var boolean Whether secure file upload is enabled or not. If enabled dangerous mime types and extensions will
      * be rejected and the file mime type needs to be verified by phps `fileinfo` extension.
      */
@@ -619,51 +627,10 @@ abstract class BaseFileSystemStorage extends Component
                 throw new Exception("Unable to find the file with id '{$fileId}', image can not be created.");
             }
 
-            $fileName = $filterId.'_'.$fileQuery->systemFileName;
-            //$fileSavePath = $this->fileServerPath($fileName);
-            
-            $tempFile = tempnam(sys_get_temp_dir(), 'prefix');
-            $tempFile.= $fileName;
+            $model = $this->createImage($fileId, $filterId);
 
-            if (empty($filterId)) {
-                $save = @copy($fileQuery->serverSource, $tempFile);
-            } else {
-                $model = StorageFilter::find()->where(['id' => $filterId])->one();
-
-                if (!$model) {
-                    throw new Exception("Could not find the provided filter id '$filterId'.");
-                }
-
-                if (!$model->applyFilterChain($fileQuery, $tempFile)) {
-                    throw new Exception("Unable to create and save image '".$tempFile."'.");
-                }
-            }
-
-            $resolution = Storage::getImageResolution($tempFile);
-            // now copy the file to the storage system
-            $this->fileSystemSaveFile($tempFile, $fileName);
-            unlink($tempFile);
-
-            // ensure the existing of the model
-            $model = StorageImage::find()->where(['file_id' => $fileId, 'filter_id' => $filterId])->one();
-
-            if ($model) {
-                $model->updateAttributes([
-                    'resolution_width' => $resolution['width'],
-                    'resolution_height' => $resolution['height'],
-                ]);
-            } else {
-                $model = new StorageImage();
-                $model->setAttributes([
-                    'file_id' => $fileId,
-                    'filter_id' => $filterId,
-                    'resolution_width' => $resolution['width'],
-                    'resolution_height' => $resolution['height'],
-                ]);
-
-                if (!$model->save()) {
-                    throw new Exception("Unable to save storage image, fatal database exception.");
-                }
+            if (!$model) {
+                throw new Exception("Unable to create the image on the filesystem.");
             }
 
             $this->_imagesArray[$model->id] = $model->toArray();
@@ -679,6 +646,73 @@ abstract class BaseFileSystemStorage extends Component
         return false;
     }
 
+    /**
+     * Just creating the image based on input informations without usage of storage files or images list.
+     * 
+     * @since 1.2.2.1
+     * @return \luya\admin\models\StorageImage|false Returns the storage image model on success, otherwise false.
+     */
+    public function createImage($fileId, $filterId)
+    {
+        $image = StorageImage::find()->where(['file_id' => $fileId, 'filter_id' => $filterId])->one();
+
+        // the image exists already in the database and the file system
+        if ($image && $image->fileExists) {
+            return $image;
+        }
+
+        $file = StorageFile::findOne($fileId);
+
+        if (!$file) {
+            return false;
+        }
+
+        // create the new image name
+        $fileName = $filterId.'_'.$file->name_new_compound;
+
+        // create a temp file
+        $tempFile = tempnam(sys_get_temp_dir(), 'prefix');
+        $tempFile.= $fileName;
+
+        // there is no filter, which means we create an image version for a given file.
+        if (empty($filterId)) {
+            @copy($file->serverSource, $tempFile);
+        } else {
+            $filter = StorageFilter::findOne($filterId);
+
+            if (!$filter) {
+                throw new Exception("Could not find the provided filter id '$filterId'.");
+            }
+
+            if (!$filter->applyFilterChain($file->serverSource, $tempFile)) {
+                throw new Exception("Unable to create and save image '".$tempFile."'.");
+            }
+        }
+
+        $resolution = Storage::getImageResolution($tempFile);
+        // now copy the file to the storage system
+        $this->fileSystemSaveFile($tempFile, $fileName);
+        unlink($tempFile);
+
+        // ensure the existing of the model
+        if ($image) {
+            $image->resolution_height = $resolution['height'];
+            $image->resolution_width = $resolution['width'];
+            $image->save();
+
+            return $image;
+        }
+
+        $image = new StorageImage();
+        $image->file_id = $fileId;
+        $image->filter_id = $filterId;
+        $image->resolution_height = $resolution['height'];
+        $image->resolution_width = $resolution['width'];
+        $image->save();
+
+        return $image;
+    }
+
     private $_foldersArray;
 
     /**
@@ -691,7 +725,17 @@ abstract class BaseFileSystemStorage extends Component
     public function getFoldersArray()
     {
         if ($this->_foldersArray === null) {
-            $this->_foldersArray = $this->getQueryCacheHelper((new Query())->from('admin_storage_folder')->select(['id', 'name', 'parent_id', 'timestamp_create'])->where(['is_deleted' => false])->orderBy(['name' => 'ASC'])->indexBy('id'), self::CACHE_KEY_FOLDER);
+            $query = (new Query())
+                ->from('admin_storage_folder as folder')
+                ->select(['folder.id', 'name', 'parent_id', 'timestamp_create', 'COUNT(file.id) filesCount'])
+                ->where(['folder.is_deleted' => false])
+                ->orderBy(['name' => 'ASC'])
+                ->leftJoin('admin_storage_file as file', 'folder.id=file.folder_id AND file.is_deleted = 0')
+                ->groupBy(['folder.id'])
+                ->indexBy(['id']);
+
+            $this->_foldersArray = $this->getQueryCacheHelper($query, self::CACHE_KEY_FOLDER);
+
         }
 
         return $this->_foldersArray;
@@ -814,6 +858,20 @@ abstract class BaseFileSystemStorage extends Component
     }
 
     /**
+     * Get the filter id based on the identifier.
+     * 
+     * This is a short hand method as its used very often
+     *
+     * @param string $identifier
+     * @return integer
+     * @since 1.2.2.1
+     */
+    public function getFilterId($identifier)
+    {
+        return $this->getFiltersArrayItem($identifier)['id'];
+    }
+
+    /**
      * Caching helper method.
      *
      * @param \yii\db\Query $query
@@ -858,8 +916,8 @@ abstract class BaseFileSystemStorage extends Component
         foreach ($this->findFiles(['is_hidden' => false, 'is_deleted' => false]) as $file) {
             if ($file->isImage) {
                 // create tiny thumbnail
-                $this->addImage($file->id, TinyCrop::identifier());
-                $this->addImage($file->id, MediumThumbnail::identifier());
+                $this->createImage($file->id, $this->getFilterId(TinyCrop::identifier()));
+                $this->createImage($file->id, $this->getFilterId(MediumThumbnail::identifier()));
             }
         }